+++ CasaEngine/Front End/Screen/MenuScreen.cs
@@ -184,7 +184,7 @@ namespace CasaEngine.FrontEnd.Screen
             Vector2 titlePosition = new Vector2(426, 80);
             Vector2 titleOrigin = font.MeasureString(menuTitle) / 2.0f;
             titlePosition = Vector2.Subtract(titlePosition, titleOrigin);
-            Color titleColor = new Color(192, 192, 192, TransitionAlpha);
+            Color titleColor = new Color((byte)192, (byte)192, (byte)192, TransitionAlpha);
             float titleScale = 1.25f;
 
             titlePosition.Y -= transitionOffset * 100;
